Move button to own Component
