+++ src/Components/About/index.js
@@ -7,23 +7,9 @@ import dispatchCenter from '../../assets/dispatchCenter.jpg';
 import careerChange from '../../assets/careerChange.jpg';
 import coding from '../../assets/coding.jpg';
 import wellsFargo from '../../assets/wellsFargo.jpg';
+import Button from '../Button';
 
 const About = () => {
-  const [classAnimate, setClassAnimate] = useState("bubbly-button");
-
-  const topFunction = () => {
-    document.body.scrollTop = 0;
-    document.documentElement.scrollTop = 0;
-  };
-
-  const animateButton = () => {
-    setClassAnimate("bubbly-button");
-    setClassAnimate((prevState) => `${prevState} animate`);
-    setTimeout(() => {
-      setClassAnimate("bubbly-button");
-      topFunction();
-    }, 500);
-  };
 
   return (
     <section className="about-container">
@@ -132,12 +118,7 @@ const About = () => {
         </p>
       </section>
       <footer className="about-footer">
-        <button
-          className={`bubbly-button ${classAnimate}`}
-          onClick={animateButton}
-        >
-          Return to top
-        </button>
+        <Button />
       </footer>
     </section>
   );
